feat: expose the utility to check if a MsgExecuteContract is a MsgTip
