+++ x/contracts/tips/types.go
@@ -1,9 +1,6 @@
 package tips
 
 import (
-	"bytes"
-	"encoding/json"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -31,34 +28,4 @@ type percentageFee struct {
 
 type fixedFee struct {
 	Amount sdk.Coins `json:"amount"`
-}
-
-// --------------------------------------------------------------------------------------------------------------------
-
-type TipMsg struct {
-	SendTip *MsgSendTip `json:"send_tip,omitempty"`
-}
-
-type MsgSendTip struct {
-	Amount sdk.Coins         `json:"amount"`
-	Target *MsgSendTipTarget `json:"target"`
-}
-
-type MsgSendTipTarget struct {
-	User    *TargetUser    `json:"user_target"`
-	Content *TargetContent `json:"content_target"`
-}
-
-func (t *MsgSendTipTarget) Equal(u *MsgSendTipTarget) bool {
-	tBz, _ := json.Marshal(t)
-	uBz, _ := json.Marshal(u)
-	return bytes.EqualFold(tBz, uBz)
-}
-
-type TargetUser struct {
-	Receiver string `json:"receiver"`
-}
-
-type TargetContent struct {
-	PostID string `json:"post_id"`
 }
